hopper updates
* restart sim using eta_abc from pandas frame
* option to refine Fhkl internally in diffBragg (for spot overlaps)
* new callback for Data modelers that saves periodicall (hopper_save_freq)
* save_up method moved from command_line/hopper.py to DataModeler class
* refine using strong_mask only (optionally dilate that mask)
* changes to nanoBragg_crystal.py: make cb_op a class member
* changes to sim_data.py: set the hall_symbol in diffBragg; sanity check on eta_abc
